Move the custom panic handler to the SDK
