+++ apps/test/app/src/main.rs
@@ -1,8 +1,5 @@
 #![cfg_attr(target_arch = "riscv32", no_std, no_main)]
 
-#[cfg(target_arch = "riscv32")]
-use sdk::fatal;
-
 extern crate alloc;
 
 mod commands;
@@ -15,22 +12,6 @@ use handlers::*;
 #[used]
 #[no_mangle]
 pub static mut APP_NAME: [u8; 32] = *b"Test\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00";
-
-#[cfg(target_arch = "riscv32")]
-#[panic_handler]
-fn my_panic(info: &core::panic::PanicInfo) -> ! {
-    let message = if let Some(location) = info.location() {
-        alloc::format!(
-            "Panic occurred in file '{}' at line {}: {}",
-            location.file(),
-            location.line(),
-            info.message()
-        )
-    } else {
-        alloc::format!("Panic occurred: {}", info.message())
-    };
-    fatal(&message); // does not return
-}
 
 #[cfg(target_arch = "riscv32")]
 #[no_mangle]
